Record of contract extension, in development.

+++ src/main/java/gmoldes/components/contract/new_contract/components/ContractParameters.java
@@ -3,6 +3,6 @@ package gmoldes.components.contract.new_contract.components;
 public class ContractParameters {
 
     public static final Integer MAXIMUM_NUMBER_OF_EXTENSIONS_OF_A_CONTRACT = 1;
-    public static final Integer NUMBER_MONTHS_MAXIMUM_DURATION_INITIAL_CONTRACT_PLUS_ITS_EXTENSIONS = 12;
+    public static final Long NUMBER_MONTHS_MAXIMUM_DURATION_INITIAL_CONTRACT_PLUS_ITS_EXTENSIONS = 12L;
 
 }
